Refactoring of utility functions to separate file
